色機能をスケジュールコンポーネントに入れました
